add criticisms/ as a separate directory
