+++ docs/source/edward.criticisms.evaluate.rst
@@ -1,0 +1,7 @@
+edward.criticisms.evaluate module
+=================================
+
+.. automodule:: edward.criticisms.evaluate
+    :members:
+    :undoc-members:
+    :show-inheritance:
