+++ docs/source/edward.criticisms.ppc.rst
@@ -1,0 +1,7 @@
+edward.criticisms.ppc module
+============================
+
+.. automodule:: edward.criticisms.ppc
+    :members:
+    :undoc-members:
+    :show-inheritance:
