+++ docs/source/edward.criticisms.rst
@@ -1,7 +1,16 @@
-edward.criticisms module
-========================
+edward.criticisms package
+=========================
 
 .. automodule:: edward.criticisms
     :members:
     :undoc-members:
     :show-inheritance:
+
+Submodules
+----------
+
+.. toctree::
+
+   edward.criticisms.evaluate
+   edward.criticisms.ppc
+
