+++ edward/criticisms/__init__.py
@@ -1,0 +1,6 @@
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+
+from edward.criticisms.evaluate import *
+from edward.criticisms.ppc import *
